+++ session.go
@@ -7,7 +7,6 @@ import (
 	"sync"
 	"time"
 
-	"github.com/cryptogarageinc/quickfix-go/config"
 	"github.com/cryptogarageinc/quickfix-go/datadictionary"
 	"github.com/cryptogarageinc/quickfix-go/internal"
 )
@@ -755,45 +754,6 @@ func (s *session) run() {
 	} else {
 		s.stopTime = time.Now().UTC()
 	}
-}
-
-func setMessageSettings(settings *SessionSettings, sessionSetting *internal.SessionSettings, timestampPrecision *TimestampPrecision) (err error) {
-	if settings.HasSetting(config.EnableLastMsgSeqNumProcessed) {
-		if sessionSetting.EnableLastMsgSeqNumProcessed, err = settings.BoolSetting(config.EnableLastMsgSeqNumProcessed); err != nil {
-			return
-		}
-	}
-
-	if settings.HasSetting(config.PersistMessages) {
-		var persistMessages bool
-		if persistMessages, err = settings.BoolSetting(config.PersistMessages); err != nil {
-			return
-		}
-
-		sessionSetting.DisableMessagePersist = !persistMessages
-	}
-	if settings.HasSetting(config.TimeStampPrecision) {
-		var precisionStr string
-		if precisionStr, err = settings.Setting(config.TimeStampPrecision); err != nil {
-			return
-		}
-
-		switch precisionStr {
-		case "SECONDS":
-			*timestampPrecision = Seconds
-		case "MILLIS":
-			*timestampPrecision = Millis
-		case "MICROS":
-			*timestampPrecision = Micros
-		case "NANOS":
-			*timestampPrecision = Nanos
-
-		default:
-			err = IncorrectFormatForSetting{Setting: config.TimeStampPrecision, Value: precisionStr}
-			return
-		}
-	}
-	return
 }
 
 // append API ------------------------------------------------------------------
